with file preview

+++ src/pages/Institution/components/ItemSettingsCard.tsx
@@ -10,14 +10,14 @@ import Button from "@material-ui/core/Button";
 import { AlertItemDelete } from "../components/AlertItemDelete";
 import { MediaItem } from "../../../api/app";
 
-function getThumbnailUrl(filePath: string) {
-  const file = filePath.replace(/^\/wikipedia\/commons\/.\/..\//i, "");
-  const filePathWithThumb = filePath.replace(
-    /^\/wikipedia\/commons\//i,
-    "/wikipedia/commons/thumb/"
-  );
-  return `https://upload.wikimedia.org${filePathWithThumb}/256px-${file}`;
-}
+// function getThumbnailUrl(filePath: string) {
+//   const file = filePath.replace(/^\/wikipedia\/commons\/.\/..\//i, "");
+//   const filePathWithThumb = filePath.replace(
+//     /^\/wikipedia\/commons\//i,
+//     "/wikipedia/commons/thumb/"
+//   );
+//   return `https://upload.wikimedia.org${filePathWithThumb}/256px-${file}`;
+// }
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -45,58 +45,67 @@ const useStyles = makeStyles((theme: Theme) =>
 
 interface ItemSettingsCardProps {
   item: MediaItem;
-  onDelete: (item: MediaItem) => void;
+  onDelete?: (item: MediaItem) => void;
+  preview?: boolean;
 }
 
-export const ItemSettingsCard = ({ item, onDelete }: ItemSettingsCardProps) => {
+export const ItemSettingsCard = ({
+  item,
+  onDelete,
+  preview,
+}: ItemSettingsCardProps) => {
+  console.log("item", item);
   const [deleteDialogOpen, setDeleteDialogOpen] = React.useState(false);
   const classes = useStyles();
   const handleItemDelete = () => {
     setDeleteDialogOpen(false);
-    onDelete(item);
+    if (onDelete) {
+      onDelete(item);
+    }
   };
   return (
     <>
       <Card key={item.filePath} className={classes.root}>
         <CardMedia
           className={classes.cover}
-          image={getThumbnailUrl(item.filePath)}
-          title="Live from space album cover"
+          image={item.thumbnailURL}
+          title={item.name}
         />
         <div className={classes.details}>
           <CardContent className={classes.content}>
-            <Typography component="h5" variant="h5">
+            <Typography component="h6" variant="h6">
               {item.name}
             </Typography>
-            <Typography variant="subtitle1" color="textSecondary">
-              {item.descriptionHTML}
-            </Typography>
           </CardContent>
-          <CardActions>
-            <Button
-              component={Link}
-              to={`/inst/met/file/${encodeURIComponent(item.filePath)}`}
-              size="small"
-              color="primary"
-            >
-              Stats
-            </Button>
-            <Button
-              size="small"
-              variant="outlined"
-              color="secondary"
-              onClick={() => setDeleteDialogOpen(true)}
-            >
-              Remove
-            </Button>
-          </CardActions>
+          {!preview && (
+            <CardActions>
+              <Button
+                component={Link}
+                to={`/inst/met/file/${encodeURIComponent(item.filePath)}`}
+                size="small"
+                color="primary"
+              >
+                Stats
+              </Button>
+              <Button
+                size="small"
+                variant="outlined"
+                color="secondary"
+                onClick={() => setDeleteDialogOpen(true)}
+              >
+                Remove
+              </Button>
+            </CardActions>
+          )}
         </div>
       </Card>
-      <AlertItemDelete
-        open={deleteDialogOpen}
-        onClose={() => setDeleteDialogOpen(false)}
-        onAgree={handleItemDelete}
-      />
+      {!preview && (
+        <AlertItemDelete
+          open={deleteDialogOpen}
+          onClose={() => setDeleteDialogOpen(false)}
+          onAgree={handleItemDelete}
+        />
+      )}
     </>
   );
 };
